Reorg of flow tests. Moving from rmutil to RLTest. Enable replication and aof test in CI.
