+++ .gitignore
@@ -3,6 +3,7 @@
 *.xml
 *.pyc
 *.log
+**/logs/**
 .vscode/
 __pycache__
 .cproject
