+++ Dockerfile.arm
@@ -14,7 +14,7 @@ RUN [ "cross-build-start" ]
 ADD ./ /build
 WORKDIR /build
 
-RUN ./deps/readies/bin/getpy2
+RUN ./deps/readies/bin/getpy3
 RUN ./system-setup.py
 RUN make fetch
