+++ src/unittests_parse_duplicate_policy.c
@@ -1,9 +1,9 @@
-#include <string.h>
-
 #include "consts.h"
 #include "generic_chunk.h"
 #include "minunit.h"
+#include "parse_policies.h"
 
+#include <string.h>
 
 #define DuplicatePolicyFromCString(str) DuplicatePolicyFromString(str, strlen(str))
 
@@ -21,7 +21,31 @@ MU_TEST(test_duplicate_policy_to_string) {
     mu_check(strcmp(DuplicatePolicyToString(DP_MIN), "min") == 0);
 }
 
+MU_TEST(test_ParseCompactionPolicy) {
+    SimpleCompactionRule *parsed_rules_out = NULL;
+    uint64_t rules_count;
+    // postive tests
+    int result = ParseCompactionPolicy(
+        "max:1m:1d;min:10s:1h;avg:2h:10d;avg:3d:100d", &parsed_rules_out, &rules_count);
+    mu_check(result == TRUE);
+    // negative tests
+    mu_check(rules_count == 4);
+    result = ParseCompactionPolicy("------", &parsed_rules_out, &rules_count);
+    mu_check(result == FALSE);
+    mu_check(rules_count == 0);
+    result = ParseCompactionPolicy("max", &parsed_rules_out, &rules_count);
+    mu_check(result == FALSE);
+    mu_check(rules_count == 0);
+    result = ParseCompactionPolicy("max:", &parsed_rules_out, &rules_count);
+    mu_check(result == FALSE);
+    mu_check(rules_count == 0);
+    result = ParseCompactionPolicy("max:abcdfeffas", &parsed_rules_out, &rules_count);
+    mu_check(result == FALSE);
+    mu_check(rules_count == 0);
+}
+
 MU_TEST_SUITE(parse_duplicate_policy_test_suite) {
-        MU_RUN_TEST(test_duplicate_policy_parse);
-        MU_RUN_TEST(test_duplicate_policy_to_string);
+    MU_RUN_TEST(test_ParseCompactionPolicy);
+    MU_RUN_TEST(test_duplicate_policy_parse);
+    MU_RUN_TEST(test_duplicate_policy_to_string);
 }
