+++ src/unittests_parse_policies.c
@@ -1,58 +1,57 @@
 /*
-* Copyright 2018-2020 Redis Labs Ltd. and Contributors
-*
-* This file is available under the Redis Labs Source Available License Agreement
-*/
-#include <stdio.h>
-#include <stdlib.h>
-#include "parse_policies.h"
-#include "minunit.h"
+ * Copyright 2018-2020 Redis Labs Ltd. and Contributors
+ *
+ * This file is available under the Redis Labs Source Available License Agreement
+ */
 #include "compaction.h"
-#include "rmutil/alloc.h"
 #include "compressed_chunk.h"
 #include "gorilla.h"
+#include "minunit.h"
+#include "parse_policies.h"
 #include "tsdb.h"
 
+#include <stdio.h>
+#include <stdlib.h>
+#include "rmutil/alloc.h"
 
 MU_TEST(test_valid_policy) {
-    SimpleCompactionRule* parsedRules;
+    SimpleCompactionRule *parsedRules;
     uint64_t rulesCount;
-    int result = ParseCompactionPolicy("max:1m:1h;min:10s:10d;last:5M:10m;avg:2h:10d;avg:3d:100d", &parsedRules, &rulesCount);
-	mu_check(result == TRUE);
-	mu_check(rulesCount == 5);
+    int result = ParseCompactionPolicy(
+        "max:1m:1h;min:10s:10d;last:5M:10m;avg:2h:10d;avg:3d:100d", &parsedRules, &rulesCount);
+    mu_check(result == TRUE);
+    mu_check(rulesCount == 5);
 
     mu_check(parsedRules[0].aggType == StringAggTypeToEnum("max"));
     mu_assert_int_eq(parsedRules[0].timeBucket, 1);
 
     mu_check(parsedRules[1].aggType == StringAggTypeToEnum("min"));
-    mu_check(parsedRules[1].timeBucket == 10*1000);
+    mu_check(parsedRules[1].timeBucket == 10 * 1000);
 
     mu_check(parsedRules[2].aggType == StringAggTypeToEnum("last"));
-    mu_check(parsedRules[2].timeBucket == 5*60*1000);
+    mu_check(parsedRules[2].timeBucket == 5 * 60 * 1000);
 
     mu_check(parsedRules[3].aggType == StringAggTypeToEnum("avg"));
-    mu_check(parsedRules[3].timeBucket == 2*60*60*1000);
+    mu_check(parsedRules[3].timeBucket == 2 * 60 * 60 * 1000);
 
     mu_check(parsedRules[4].aggType == StringAggTypeToEnum("avg"));
-    mu_check(parsedRules[4].timeBucket == 3*60*60*1000*24);
+    mu_check(parsedRules[4].timeBucket == 3 * 60 * 60 * 1000 * 24);
     free(parsedRules);
 }
 
-
 MU_TEST(test_invalid_policy) {
-    SimpleCompactionRule* parsedRules;
+    SimpleCompactionRule *parsedRules;
     uint64_t rulesCount;
     int result;
     result = ParseCompactionPolicy("max:1M;mins:10s;avg:2h;avg:1d", &parsedRules, &rulesCount);
-	mu_check(result == FALSE);
-	mu_check(rulesCount == 0);
+    mu_check(result == FALSE);
+    mu_check(rulesCount == 0);
 
     result = ParseCompactionPolicy("max:12hd;", &parsedRules, &rulesCount);
-	mu_check(result == FALSE);
-	mu_check(rulesCount == 0);
+    mu_check(result == FALSE);
+    mu_check(rulesCount == 0);
     free(parsedRules);
 }
-
 
 MU_TEST(test_StringLenAggTypeToEnum) {
     mu_check(StringAggTypeToEnum("min") == TS_AGG_MIN);
@@ -66,7 +65,7 @@ MU_TEST(test_StringLenAggTypeToEnum) {
 }
 
 MU_TEST_SUITE(parse_policies_test_suite) {
-	MU_RUN_TEST(test_valid_policy);
-	MU_RUN_TEST(test_invalid_policy);
-	MU_RUN_TEST(test_StringLenAggTypeToEnum);
+    MU_RUN_TEST(test_valid_policy);
+    MU_RUN_TEST(test_invalid_policy);
+    MU_RUN_TEST(test_StringLenAggTypeToEnum);
 }
