+++ tests/flow/includes.py
@@ -1,6 +1,5 @@
-
-import sys
 import os
+import sys
 
 try:
     sys.path.insert(0, os.path.join(os.path.dirname(__file__), "../../deps/readies"))
