+++ tests/flow/requirements.txt
@@ -1,0 +1,6 @@
+pytest==3.0.7
+redis>=3.0.0
+git+https://github.com/Grokzen/redis-py-cluster.git@master
+git+https://github.com/RedisLabsModules/RLTest.git@master
+ramp-packer
+statistics
