--- tests/requirements.txt
@@ -1,5 +1,0 @@
-pytest==3.0.7
-redis==2.10.5
-rmtest==0.7.0
-ramp-packer
-statistics
